+++ src/frontend/js/pages/TeacherDashboardCourseLearnersLayout/TeacherDashboardLearners/index.tsx
@@ -1,5 +1,0 @@
-const TeacherDashboardLearners = () => {
-  return "i'm TeacherDashboardLearners";
-};
-
-export default TeacherDashboardLearners;